test: isolate AmpForm integration tests
