+++ tests/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Required to set mypy options for the tests folder."""
