+++ tests/integration/ampform/test_callbacks.py
