+++ tox.ini
@@ -10,11 +10,11 @@ skipsdist = True
 
 [testenv]
 description =
-    Run all unit tests
+    Run all unit tests and doctests
 allowlist_externals =
     pytest
 commands =
-    pytest {posargs}
+    pytest src {posargs:tests/unit}
 
 [testenv:cov]
 description =
@@ -116,6 +116,18 @@ allowlist_externals =
 commands =
     mypy src tests # run separately because of potential caching problems
     pre-commit run {posargs} -a
+
+[testenv:test]
+description =
+    Run unit tests, integration tests, and doctests
+allowlist_externals =
+    pytest
+commands =
+    pytest {posargs} \
+        --cov-fail-under=90 \
+        --cov-report=html \
+        --cov-report=xml \
+        --cov=qrules
 
 [testenv:testnb]
 description =
